M26 add access right

+++ app/Http/Controllers/Asset/BrandsController.php
@@ -11,6 +11,11 @@ class BrandsController extends Controller
 {
     public function index()
     {
+        if(!\Auth::user()->checkAccessById(54, 'V')) {
+            \Session::flash('error', "You don't have permission"); 
+            return redirect("/home"); 
+        }
+
         $brands = Brands::all();
         return view('assets.brands.index', [
             'brands' => $brands
@@ -18,6 +23,11 @@ class BrandsController extends Controller
     }
     public function store(Request $request)
     {
+        if(!\Auth::user()->checkAccessById(54, 'A')) {
+            \Session::flash('error', "You don't have permission"); 
+            return redirect("/home"); 
+        }
+
         Brands::create($request->only('description'));
         \Session::flash('success', 'Brand has been created successfully');
 
@@ -25,12 +35,22 @@ class BrandsController extends Controller
     }
     public function destroy($id)
     {
+        if(!\Auth::user()->checkAccessById(54, 'D')) {
+            \Session::flash('error', "You don't have permission"); 
+            return redirect("/home"); 
+        }
+
         $brand = Brands::findOrFail($id);
         $brand->delete();
         return response()->json(['success' => true]);
     }
     public function update(Request $request, $id)
     {
+        if(!\Auth::user()->checkAccessById(54, 'E')) {
+            \Session::flash('error', "You don't have permission"); 
+            return redirect("/home"); 
+        }
+
         $brand = Brands::findOrFail($id);
         $brand->update(
             $request->only([
